+++ qmule/src/org/qcmg/qmule/io/Dbsnp130Record.java
@@ -9,7 +9,7 @@
 //
 
 
-package org.qcmg.dbsnp;
+package org.qcmg.qmule.io;
