+++ qmule/src/org/qcmg/qmule/io/VerifiedSnpRecord.java
@@ -1,7 +1,7 @@
 /**
  * © Copyright The University of Queensland 2010-2014.  This code is released under the terms outlined in the included LICENSE file.
  */
-package org.qcmg.pileup;
+package org.qcmg.qmule.io;
 
 import org.qcmg.common.model.Classification;
