+++ qmule/src/org/qcmg/qmule/snppicker/ExamineVerifiedSnps.java
@@ -21,8 +21,8 @@ import org.qcmg.common.vcf.VcfRecord;
 import org.qcmg.common.vcf.VcfUtils;
 import org.qcmg.pileup.QPileupFileReader;
 import org.qcmg.pileup.QSnpRecord;
-import org.qcmg.pileup.VerifiedSnpFileReader;
-import org.qcmg.pileup.VerifiedSnpRecord;
+import org.qcmg.qmule.io.VerifiedSnpFileReader;
+import org.qcmg.qmule.io.VerifiedSnpRecord;
 import org.qcmg.vcf.VCFFileReader;
 
 public class ExamineVerifiedSnps {
