+++ qmule/src/org/qcmg/qmule/snppicker/SnpPicker.java
@@ -16,8 +16,6 @@ import java.util.TreeMap;
 
 import htsjdk.samtools.SAMRecord;
 
-import org.qcmg.chrconv.ChrConvFileReader;
-import org.qcmg.chrconv.ChromosomeConversionRecord;
 import org.qcmg.common.log.QLogger;
 import org.qcmg.common.log.QLoggerFactory;
 import org.qcmg.common.model.ChrPointPosition;
@@ -28,8 +26,6 @@ import org.qcmg.common.util.FileUtils;
 import org.qcmg.common.util.TabTokenizer;
 import org.qcmg.common.vcf.VcfRecord;
 import org.qcmg.common.vcf.VcfUtils;
-import org.qcmg.dbsnp.Dbsnp130Record;
-import org.qcmg.dbsnp.DbsnpFileReader;
 import org.qcmg.gff3.GFF3FileReader;
 import org.qcmg.gff3.GFF3Record;
 import org.qcmg.illumina.IlluminaFileReader;
@@ -39,6 +35,10 @@ import org.qcmg.pileup.PileupFileReader;
 import org.qcmg.qmule.Messages;
 import org.qcmg.qmule.Options;
 import org.qcmg.qmule.QMuleException;
+import org.qcmg.qmule.io.ChrConvFileReader;
+import org.qcmg.qmule.io.ChromosomeConversionRecord;
+import org.qcmg.qmule.io.Dbsnp130Record;
+import org.qcmg.qmule.io.DbsnpFileReader;
 import org.qcmg.record.Record;
 import org.qcmg.vcf.VCFFileReader;
